+++ InvokeBuildHelper/Scripts/ScriptAnalyzerTests.ps1
@@ -42,27 +42,48 @@ Describe 'Script Analyzer' {
     {
         Context "$($currentRule.RuleName) ($($currentRule.Severity))" {
 
-            $currentIssues = $issues.Where({ $_.RuleName -eq $currentRule.RuleName })
-            $currentIssues | ForEach-Object { $_ | Add-Member -MemberType 'NoteProperty' -Name 'RelativePath' -Value ('\' + $_.ScriptPath.Replace($BuildRoot, '').TrimStart('\')) }
-            $currentIssues = $currentIssues | Where-Object { $relativePath = $_.RelativePath; @($ExcludePath | Where-Object { $relativePath -like $_ }).Count -eq 0 }
+            $currentIssues =
+                $issues |
+                    Where-Object { $_.RuleName -eq $currentRule.RuleName } |
+                        ForEach-Object { @{ RelativePath = '\' + $_.ScriptPath.Replace($BuildRoot, '').TrimStart('\'); Line = $_.Line; Column = $_.Column; Message = $_.Message } } |
+                            Where-Object { $relativePath = $_.RelativePath; @($ExcludePath | Where-Object { $relativePath -like $_ }).Count -eq 0 }
 
-            if ($currentIssues.Count -eq 0)
+            if (@($currentIssues).Count -eq 0)
             {
                 It 'Should pass the rule for all files' {
 
-                    $currentIssues.Count | Should -Be 0
+                    0 | Should -Be 0
                 }
             }
             else
             {
-                foreach ($currentIssue in $currentIssues)
-                {
-                    It ('Should pass the rule for {0}:{1},{2}' -f $currentIssue.RelativePath, $currentIssue.Line, $currentIssue.Column) {
+                It 'Should pass the rule for <RelativePath>:<Line>,<Column>' -ForEach $currentIssues {
 
-                        throw $currentIssue.Message
-                    }
+                    param ($Message)
+
+                    # Assert
+                    throw $Message
                 }
             }
+
+            # $currentIssueCount = @($currentIssues).Count
+
+            # if ($currentIssueCount -eq 0)
+            # {
+            #     It 'Should pass the rule for all files' {
+
+            #         0 | Should -Be 0
+            #     }
+            # }
+            # else
+            # {
+            #     foreach ($currentIssue in $currentIssues)
+            #     {
+            #         It ('Should pass the rule for {0}:{1},{2}' -f $currentIssue.RelativePath, $currentIssue.Line, $currentIssue.Column) {
+
+            #         }
+            #     }
+            # }
         }
     }
 }