+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is mainly based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+* Added: Support for faster dot-sourcing module loading and module debugging mode
+* Added: PowerShell module script (psm1) tests for module behavior, metadata and loader
+* Added: Additional checks for the module helper and function files
+
 ## 3.3.4 - 2026-02-14
 
 * Fixed: Fix the DSC Resource schema tests when using Pester v5